Refactor Views Into Their Own File
